# openapi.json
{
  "openapi": "3.0.1",
  "info": { "title": "OpenAPI definition", "version": "v0" },
  "servers": [
    {
      "url": "http://localhost:8080/api",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/auth/register": {
      "post": {
        "tags": ["authentication-controller"],
        "operationId": "registerUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SignupRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": { "*/*": { "schema": { "type": "string" } } }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["authentication-controller"],
        "summary": "Authenticate user",
        "description": "Authenticate user",
        "operationId": "authenticateUser",
        "requestBody": {
          "description": "Login request",
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/JwtResponse" }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": { "schema": { "type": "string" } }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": { "schema": { "type": "string" } }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": { "type": "string" },
                "example": "error message"
              }
            }
          }
        }
      }
    },
...
    "/images": {
      "get": {
        "tags": ["image-controller"],
        "summary": "Get all images",
        "description": "Get all images in the upload directory",
        "operationId": "getImages",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ImageMetadata" }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": { "schema": { "type": "string" } }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "SignupRequest": {
        "type": "object",
        "properties": {
          "username": { "type": "string" },
          "email": { "type": "string" },
          "password": { "type": "string" }
        }
      },
      "LoginRequest": {
        "type": "object",
        "properties": {
          "username": { "type": "string" },
          "password": { "type": "string" }
        }
      },
      "JwtResponse": {
        "type": "object",
        "properties": {
          "token": { "type": "string" },
          "id": { "type": "integer", "format": "int64" },
          "username": { "type": "string" },
          "email": { "type": "string" },
          "roles": {
            "uniqueItems": true,
            "type": "array",
            "items": { "type": "string" }
          },
          "type": { "type": "string" }
        }
      },
      "ImageMetadata": {
        "type": "object",
        "properties": {
          "fileName": { "type": "string" },
          "alt": { "type": "string" },
          "width": { "type": "integer", "format": "int32" },
          "height": { "type": "integer", "format": "int32" }
        }
      }
    }
  }
}